refactor: consolidate sign message routing into single component
Merged SignMessageRoute logic directly into SignMessage component
and extracted the view logic into SignMessageView for better
separation of concerns and storybook compatibility.
